Старт работы по созданию карточек
